This PR fixes some ASAN and UBSAN findings across the cerberus-core code. Most of the findings are buffer overflows within unit test code not affecting the logic, but some are also in the core code itself.

----
#### AI description  (iteration 1)
#### PR Classification
Bug fix addressing memory safety issues by replacing unsafe string functions and ensuring proper size calculations.

#### PR Summary
This pull request fixes address sanitizer findings by replacing dangerous usages of strcpy with memcpy (using explicit length calculations) and adding proper type casts and bounds checks across the codebase and tests.
- `core/spdm/spdm_commands.c`: Replaced several instances of strcpy with memcpy and introduced a local variable for context length.
- `core/testing/spdm/*` and `core/testing/attestation/*`: Updated multiple tests to use memcpy with correct buffer sizes and fixed mock argument indices.
- `core/common/common_math.c`, `firmware_update_handler.c`, and `RiotDerEnc.c`: Added explicit casts and corrected bit shifts and status updates.
- `core/testing/asn1/*`: Standardized memcpy length usage in DER encoding/decoding tests by using defined constant lengths instead of sizeof on raw buffers.
- `core/testing/cmd_interface/*` and `core/testing/flash/*`: Adjusted memcpy parameters and output size checks to ensure accurate memory copying and prevent buffer overflows.
<!-- GitOpsUserAgent=GitOps.Apps.Server.pullrequestcopilot -->

Related work items: #5504

+++ core/cmd_interface/cerberus_protocol_required_commands.c
@@ -57,7 +57,7 @@ void cerberus_protocol_build_error_response (struct cmd_interface_msg *message, 
 	if (error_code != CERBERUS_PROTOCOL_NO_ERROR) {
 		debug_log_create_entry (DEBUG_LOG_SEVERITY_ERROR, DEBUG_LOG_COMPONENT_CMD_INTERFACE,
 			CMD_LOGGING_CERBERUS_REQUEST_FAIL,
-			((error_code << 24) | (command_code << 16) | (message->source_eid << 8) |
+			(((uint32_t) error_code << 24) | (command_code << 16) | (message->source_eid << 8) |
 					message->channel_id), error_data);
 	}
 }
